refactoring tests and avoid doing it twice.
